950: Odešleme info pro CFO že se položky měly přestat prodávat
