+++ admin/cron/_cron_na_cas.php
@@ -19,6 +19,10 @@ if ($sleep >= 300) {
     return false;
 }
 
+if ($sleep <= 0) {
+    return true; // možná nějaký job běžel před tímto a zdržel další v pořadí (z nouze o zdroje je pouštíme jednovláknově)
+}
+
 set_time_limit($sleep + 1);
 
 sleep($sleep);
