+++ admin/cron/mail_cfo_brzke_odhlaseni_neplaticu.php
@@ -6,6 +6,7 @@ use Gamecon\Cas\DateTimeCz;
 use Gamecon\Uzivatel\Exceptions\NevhodnyCasProHromadneOdhlasovani;
 use Gamecon\Cas\DateTimeGamecon;
 use Gamecon\Report\BfgrReport;
+use Gamecon\Shop\Shop;
 
 require_once __DIR__ . '/_cron_zavadec.php';
 
@@ -52,15 +53,42 @@ if (!$poradiOznameni) {
 // abychom měli čerstvé informace o neplatičích
 require __DIR__ . '/fio_stazeni_novych_plateb.php';
 
-$zpravy = [];
+$zpravyNeplatici = [];
 try {
     foreach ($hromadneOdhlaseniNeplaticu->neplaticiAKategorie()
              as ['uzivatel' => $uzivatel, 'kategorie_neplatice' => $kategorieNeplatice]) {
         /** @var \Gamecon\Uzivatel\KategorieNeplatice $kategorieNeplatice */
-        $zpravy[] = "Účastník '{$uzivatel->jmenoNick()}' ({$uzivatel->id()}) bude zítra odhlášen, protože má kategorii neplatiče {$kategorieNeplatice->ciselnaKategoriiNeplatice()}";
+        $zpravyNeplatici[] = "Účastník '{$uzivatel->jmenoNick()}' ({$uzivatel->id()}) bude zítra odhlášen, protože má kategorii neplatiče {$kategorieNeplatice->ciselnaKategoriiNeplatice()}";
     }
 } catch (NevhodnyCasProHromadneOdhlasovani $nevhodnyCasProHromadneOdhlasovani) {
     return;
+}
+
+$zpravyPolozky = [];
+foreach (Shop::letosniPolozkySeSpatnymKoncem($systemoveNastaveni) as $polozkaSeSpatnymKoncem) {
+    $nabizetDoDleNastaveni = $polozkaSeSpatnymKoncem->doKdyNabizetDleNastaveni($systemoveNastaveni);
+    if (!$nabizetDoDleNastaveni || !$polozkaSeSpatnymKoncem->nabizetDo()) {
+        trigger_error(
+            "Polozka '{$polozkaSeSpatnymKoncem->nazev()}' ({$polozkaSeSpatnymKoncem->idPredmetu()}) je údajně se špatným koncem, ale nemá žádný konec prodeje",
+            E_USER_WARNING
+        );
+        continue;
+    }
+    if ($nabizetDoDleNastaveni->getTimestamp() === $polozkaSeSpatnymKoncem->nabizetDo()->getTimestamp()) {
+        trigger_error(
+            "Polozka '{$polozkaSeSpatnymKoncem->nazev()}' ({$polozkaSeSpatnymKoncem->idPredmetu()}) je údajně se špatným koncem, ale přitom má konec prodeje správně",
+            E_USER_WARNING
+        );
+        continue;
+    }
+    if ($nabizetDoDleNastaveni->getTimestamp() > $polozkaSeSpatnymKoncem->nabizetDo()->getTimestamp()) {
+        continue; // že se přestala prodávat o něco dřív nás teď nezajímá
+    }
+    if ($polozkaSeSpatnymKoncem->nabizetDo() < $systemoveNastaveni->ted()) {
+        continue; // pokud se už neprodává, tak nás teď nezajímá
+    }
+
+    $zpravyPolozky[] = "Položka '{$polozkaSeSpatnymKoncem->nazev()}' ({$polozkaSeSpatnymKoncem->idPredmetu()}) by se měla přestat podávat {$nabizetDoDleNastaveni->formatCasStandard()}, ale bude se prodávat až do {$polozkaSeSpatnymKoncem->nabizetDo()->formatCasStandard()}. Opravte její datum konce prodeje v shopu.";
 }
 
 $bfgrSoubor = sys_get_temp_dir() . '/' . uniqid('bfgr-', true) . '.xlsx';
@@ -68,8 +96,7 @@ $bfgrReport = new BfgrReport($systemoveNastaveni);
 $bfgrReport->exportuj('xlsx', true, $bfgrSoubor);
 
 $cfosEmaily    = Uzivatel::cfosEmaily();
-$budeOdhlaseno = count($zpravy);
-$zpravyString  = implode(";\n", $zpravy);
+$budeOdhlaseno = count($zpravyNeplatici);
 $brzy          = match ($poradiOznameni) {
     1 => 'Zítra',
     2 => 'Za hodinu',
@@ -77,6 +104,18 @@ $brzy          = match ($poradiOznameni) {
 };
 $uvod          = "$brzy Gamecon systém odhlásí $budeOdhlaseno účastníků z letošního Gameconu, protože jsou neplatiči.";
 $oddelovac     = str_repeat('═', mb_strlen($uvod));
+$zpravyString  = implode(";\n", $zpravyNeplatici);
+
+if ($zpravyPolozky) {
+    $zpravyPolozkyString = implode(";\n", $zpravyPolozky);
+    $zpravyString        .= <<<TEXT
+
+$oddelovac
+
+$zpravyPolozkyString
+TEXT;
+}
+
 (new GcMail())
     ->adresati($cfosEmaily ?: ['info@gamecon.cz'])
     ->predmet("$brzy bude hromadně odhlášeno $budeOdhlaseno neplatičů z GC")
