+++ model/Shop/Polozka.php
@@ -2,34 +2,43 @@
 
 namespace Gamecon\Shop;
 
-use Gamecon\Cas\DateTimeCz;
+use Gamecon\Cas\DateTimeImmutableStrict;
+use Gamecon\SystemoveNastaveni\SystemoveNastaveni;
 
 class Polozka
 {
-    private $idPredmetu;
-    private $nazev;
-    private $cena;
-    private $suma;
-    private $modelRok;
-    private $naposledyKoupenoKdy;
-    private $prodanoKusu;
-    private $vyrobenoKusu;
-    private $zbyvaKusu;
-    private $idTypu;
+    private int $idPredmetu;
+    private string $nazev;
+    private float $cena;
+    private float $suma;
+    private int $modelRok;
+    private ?DateTimeImmutableStrict $naposledyKoupenoKdy;
+    private float $prodanoKusu;
+    private float $vyrobenoKusu;
+    private ?int $kategoriePredmetu;
+    private ?DateTimeImmutableStrict $nabizetDo;
+    private float $zbyvaKusu;
+    private int $idTypu;
 
     public function __construct(array $hodnoty) {
-        $this->idPredmetu = (int)$hodnoty['id_predmetu'];
-        $this->nazev = (string)$hodnoty['nazev'];
-        $this->cena = (float)$hodnoty['cena_aktualni'];
-        $this->suma = (float)$hodnoty['suma'];
-        $this->modelRok = (int)$hodnoty['model_rok'];
+        $this->idPredmetu          = (int)$hodnoty['id_predmetu'];
+        $this->nazev               = (string)$hodnoty['nazev'];
+        $this->cena                = (float)$hodnoty['cena_aktualni'];
+        $this->suma                = (float)$hodnoty['suma'];
+        $this->modelRok            = (int)$hodnoty['model_rok'];
         $this->naposledyKoupenoKdy = $hodnoty['naposledy_koupeno_kdy']
-            ? new DateTimeCz($hodnoty['naposledy_koupeno_kdy'])
+            ? new DateTimeImmutableStrict($hodnoty['naposledy_koupeno_kdy'])
             : null;
-        $this->prodanoKusu = (float)$hodnoty['prodano_kusu'];
-        $this->vyrobenoKusu = (float)$hodnoty['kusu_vyrobeno'];
-        $this->zbyvaKusu = $this->vyrobenoKusu - $this->prodanoKusu;
-        $this->idTypu = (int)$hodnoty['typ'];
+        $this->prodanoKusu         = (float)$hodnoty['prodano_kusu'];
+        $this->vyrobenoKusu        = (float)$hodnoty['kusu_vyrobeno'];
+        $this->kategoriePredmetu   = $hodnoty['kategorie_predmetu']
+            ? (int)$hodnoty['kategorie_predmetu']
+            : null;
+        $this->nabizetDo           = $hodnoty['nabizet_do']
+            ? new DateTimeImmutableStrict($hodnoty['nabizet_do'])
+            : null;
+        $this->zbyvaKusu           = $this->vyrobenoKusu - $this->prodanoKusu;
+        $this->idTypu              = (int)$hodnoty['typ'];
     }
 
     public function idPredmetu(): ?int {
@@ -56,7 +65,7 @@ class Polozka
         return $this->prodanoKusu;
     }
 
-    public function naposledyKoupenoKdy(): ?DateTimeCz {
+    public function naposledyKoupenoKdy(): ?DateTimeImmutableStrict {
         return $this->naposledyKoupenoKdy;
     }
 
@@ -71,4 +80,22 @@ class Polozka
     public function idTypu(): int {
         return $this->idTypu;
     }
+
+    public function kategoriePredmetu(): ?int {
+        return $this->kategoriePredmetu;
+    }
+
+    public function nabizetDo(): ?DateTimeImmutableStrict {
+        return $this->nabizetDo;
+    }
+
+    public function doKdyNabizetDleNastaveni(SystemoveNastaveni $systemoveNastaveni): ?DateTimeImmutableStrict {
+        return match ($this->idTypu()) {
+            TypPredmetu::JIDLO => $systemoveNastaveni->prodejJidlaDo(),
+            TypPredmetu::TRICKO => $systemoveNastaveni->prodejTricekDo(),
+            TypPredmetu::PREDMET => $systemoveNastaveni->prodejPredmetuBezTricekDo(),
+            default => null,
+        };
+    }
+
 }
